Move processors into their own package.
